pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-guard"
version = "0.1.0"
description = "A library for masking sensitive data in RAG applications"
readme = "README.md"
authors = [{name = "engchina"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=0.19.0"
]

[project.urls]
Homepage = "https://github.com/example/rag-guard"
Repository = "https://github.com/example/rag-guard"